+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - When the `tags:` header on a node is blank, `Node.Tags` now returns an empty collection, instead of a collection containing a single empty string.
 - Fixed an issue that would cause testplans with syntax errors to incorrectly pass.
 - Fixed an issue where using `visited()` or `visited_count()` would not correctly track visiting node groups.
+- Language Server: NodeInfo objects now indicate whether a node contains any jumps to nodes in a different file.
 
 ### Removed